requirements.txt
Django==3.1.7
django-heroku==0.3.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
rehive==1.2.16
requests==2.25.1
sentry-sdk==1.0.0
django-healthz==0.0.5
django-rehive-extras==0.0.13
drf-rehive-extras==0.0.13
django-storages[google]==1.11.1
stellar-sdk==3.1.3
google-cloud-kms==2.2.0
crcmod==1.7
